refactor: clean up zone dev implementation
