+++ Justfile
@@ -27,24 +27,19 @@ build binary extra_args="":
 regen-zone-dev-genesis:
     #!/bin/bash
     set -euo pipefail
-    TEMPO_GENESIS_HEADER_RLP="f90217808080f90211a00000000000000000000000000000000000000000000000000000000000000000a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421940000000000000000000000000000000000000000a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421a056e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421b9010000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000808080808080a0000000000000000000000000000000000000000000000000000000000000000088000000000000000080a000000000000000000000000000000000000000000000000000000000000000008080"
     rm -rf {{zone_dev_genesis_tmp}}
     forge build --root specs/ref-impls --no-lint
     cargo run -p tempo-xtask -- generate-zone-genesis \
         --output {{zone_dev_genesis_tmp}} \
         --chain-id 1337 \
         --tempo-portal 0x0000000000000000000000000000000000000000 \
-        --tempo-genesis-header-rlp "$TEMPO_GENESIS_HEADER_RLP" \
         --admin 0xaAaAaAaa00000000000000000000000000000000 \
         --specs-out specs/ref-impls/out \
         --with-createx \
         --with-safe-deployer \
         --with-create2-factory \
         --with-zone-factory-bytecode
-    tmp="$(mktemp)"
-    jq '.alloc["0x1c00000000000000000000000000000000000000"].storage["0x0000000000000000000000000000000000000000000000000000000000000000"] = "0xb049644b1d5a0ec9d785dd48f95099e0f566112084acb1ba0814112209b432a1"' \
-        {{zone_dev_genesis_tmp}}/genesis.json > "$tmp"
-    mv "$tmp" crates/node/assets/zone-dev-genesis.json
+    mv {{zone_dev_genesis_tmp}}/genesis.json crates/node/assets/zone-dev-genesis.json
     rm -rf {{zone_dev_genesis_tmp}}
 
 [group('localnet')]
