+++ crates/node/Cargo.toml
@@ -60,7 +60,6 @@ alloy-provider = { workspace = true, features = ["ws"] }
 alloy-rpc-client.workspace = true
 alloy-rpc-types-engine.workspace = true
 alloy-rpc-types-eth.workspace = true
-alloy-signer.workspace = true
 alloy-signer-local.workspace = true
 alloy-sol-types.workspace = true
