+++ crates/sequencer/Cargo.toml
@@ -26,6 +26,7 @@ alloy-provider = { workspace = true, features = ["ws"] }
 alloy-rlp.workspace = true
 alloy-rpc-client.workspace = true
 alloy-rpc-types-eth.workspace = true
+alloy-signer.workspace = true
 alloy-signer-local.workspace = true
 alloy-sol-types.workspace = true
 alloy-transport.workspace = true
@@ -33,6 +34,7 @@ alloy-transport.workspace = true
 # misc
 eyre.workspace = true
 futures.workspace = true
+k256.workspace = true
 metrics.workspace = true
 parking_lot.workspace = true
 tokio.workspace = true
